only validate a required minimum amount of properties for user creation and update. Remove deprecated tests. Remove double validation in the CustomerIoUpdateCustomerMessage right before calling the C.io identify function (for upserting user in C.io)
