+++ test/unit/messages/UserMessage.test.js
@@ -4,7 +4,6 @@
 
 const test = require('ava');
 const chai = require('chai');
-const Chance = require('chance');
 
 const MessageValidationBlinkError = require('../../../src/errors/MessageValidationBlinkError');
 const UserMessage = require('../../../src/messages/UserMessage');
@@ -14,7 +13,6 @@ const MessageValidationHelper = require('../../helpers/MessageValidationHelper')
 // ------- Init ----------------------------------------------------------------
 
 chai.should();
-const chance = new Chance();
 const generator = MessageFactoryHelper.getValidUser;
 const mutator = function ({ remove, change, value, message }) {
   const mutant = message;
@@ -38,7 +36,7 @@ test('User message generator', () => {
 test('Validate a hundred fake users', () => {
   let count = 100;
   while (count > 0) {
-    generator().validateStrict.should.not.throw(MessageValidationBlinkError);
+    generator().validate.should.not.throw(MessageValidationBlinkError);
     count -= 1;
   }
 });
@@ -58,95 +56,6 @@ test('User Message should remove certain optional fields when empty', () => {
     'mobile',
   ]
     .forEach(field => MessageValidationHelper.removesWhenEmpty(field, generator, mutator));
-});
-
-test('User Message optional fields should have correct default values', () => {
-  const mapping = {
-    last_authenticated_at: null,
-    birthdate: null,
-    facebook_id: null,
-    first_name: null,
-    last_name: null,
-    addr_city: null,
-    addr_state: null,
-    addr_zip: null,
-    source: null,
-    source_detail: null,
-    language: null,
-    country: null,
-    role: 'user',
-    interests: null,
-    sms_status: null,
-  };
-  Object.entries(mapping).forEach(([field, defaultValue]) => {
-    MessageValidationHelper.defaultsToWhenEmpty(field, defaultValue, generator, mutator);
-  });
-});
-
-test('User Message should fail on incorrect types', () => {
-  const mapping = {
-    id: chance.integer(),
-    email: chance.integer(),
-    mobile: chance.integer(),
-    updated_at: chance.integer(),
-    created_at: chance.integer(),
-    // no sms_status
-    last_authenticated_at: chance.integer(),
-    birthdate: chance.integer(),
-    facebook_id: chance.integer(),
-    first_name: chance.integer(),
-    last_name: chance.integer(),
-    addr_city: chance.integer(),
-    addr_state: chance.integer(),
-    addr_zip: chance.integer(),
-    source: chance.integer(),
-    source_detail: chance.integer(),
-    language: chance.integer(),
-    country: chance.integer(),
-    role: chance.integer(),
-    interests: chance.word(),
-  };
-  Object.entries(mapping).forEach(([field, incorrectValue]) => {
-    MessageValidationHelper.ensureType(field, incorrectValue, generator, mutator);
-  });
-});
-
-test('Test mobile only users', () => {
-  let mutant;
-
-  // Test normal user.
-  generator().isMobileOnly().should.be.false;
-
-  // Test no email
-  mutant = mutator({
-    remove: 'email',
-    message: generator(),
-  });
-  mutant.isMobileOnly().should.be.true;
-
-  // Test no email
-  mutant = mutator({
-    change: 'email',
-    value: '',
-    message: generator(),
-  });
-  mutant.isMobileOnly().should.be.true;
-
-  // Test 15554443332@mobile.import
-  mutant = mutator({
-    change: 'email',
-    value: '15554443332@mobile.import',
-    message: generator(),
-  });
-  mutant.isMobileOnly().should.be.true;
-
-  // Test runscope source
-  mutant = mutator({
-    change: 'source',
-    value: 'runscope',
-    message: generator(),
-  });
-  mutant.isMobileOnly().should.be.true;
 });
 
 // ------- End -----------------------------------------------------------------
